Share the shadow-piercing selector body between the two patches
_customFindFramesByParsed (frameSelectorsPatch) and _customFindElementsByParsed
(framesPatch) inject the same ~147-line traversal. They differ in four places:
the recursive call target, how ElementHandle is reachable in the patched
module's scope, the callId expression, and the receiver that owns
_findElementPositionInDomTree.

Move the body into driver_patches/customFindByParsed.ts and parameterise those
four. Both call sites emit byte-identical text to before, so the generated
patch is unchanged.
